Breaks down SocketServer classes into ConnectionQuotaEntity.scala and ConnectionQuotas.scala
Since SocketServer is too large, breaking it down to make the classes more
manageable.
