+++ core/src/main/scala/kafka/network/SocketServer.scala
@@ -25,7 +25,6 @@ import java.util
 import java.util.concurrent._
 import java.util.concurrent.atomic._
 import kafka.cluster.{BrokerEndPoint, EndPoint}
-import kafka.network.ConnectionQuotas._
 import kafka.network.Processor._
 import kafka.network.RequestChannel.{CloseConnectionResponse, EndThrottlingResponse, NoOpResponse, SendResponse, StartThrottlingResponse}
 import kafka.network.SocketServer._
@@ -34,18 +33,17 @@ import kafka.server.{ApiVersionManager, BrokerReconfigurable, KafkaConfig}
 import org.apache.kafka.common.message.ApiMessageType.ListenerType
 import kafka.utils._
 import org.apache.kafka.common.config.ConfigException
-import org.apache.kafka.common.config.internals.QuotaConfigs
 import org.apache.kafka.common.errors.InvalidRequestException
 import org.apache.kafka.common.memory.{MemoryPool, SimpleMemoryPool}
 import org.apache.kafka.common.metrics._
-import org.apache.kafka.common.metrics.stats.{Avg, CumulativeSum, Meter, Rate}
+import org.apache.kafka.common.metrics.stats.{CumulativeSum, Meter}
 import org.apache.kafka.common.network.KafkaChannel.ChannelMuteEvent
 import org.apache.kafka.common.network.{ChannelBuilder, ChannelBuilders, ClientInformation, KafkaChannel, ListenerName, ListenerReconfigurable, NetworkSend, Selectable, Send, Selector => KSelector}
 import org.apache.kafka.common.protocol.ApiKeys
 import org.apache.kafka.common.requests.{ApiVersionsRequest, RequestContext, RequestHeader}
 import org.apache.kafka.common.security.auth.SecurityProtocol
 import org.apache.kafka.common.utils.{KafkaThread, LogContext, Time, Utils}
-import org.apache.kafka.common.{Endpoint, KafkaException, MetricName, Reconfigurable}
+import org.apache.kafka.common.{Endpoint, KafkaException, Reconfigurable}
 import org.apache.kafka.server.metrics.KafkaMetricsGroup
 import org.apache.kafka.server.util.FutureUtils
 import org.slf4j.event.Level
@@ -1378,480 +1376,6 @@ private[kafka] class Processor(
       metrics.removeMetric(expiredConnectionsKilledCountMetricName)
     }
   }
-}
-
-/**
- * Interface for connection quota configuration. Connection quotas can be configured at the
- * broker, listener or IP level.
- */
-sealed trait ConnectionQuotaEntity {
-  def sensorName: String
-  def metricName: String
-  def sensorExpiration: Long
-  def metricTags: Map[String, String]
-}
-
-object ConnectionQuotas {
-  private val InactiveSensorExpirationTimeSeconds = TimeUnit.HOURS.toSeconds(1)
-  private val ConnectionRateSensorName = "Connection-Accept-Rate"
-  private val ConnectionRateMetricName = "connection-accept-rate"
-  private val IpMetricTag = "ip"
-  private val ListenerThrottlePrefix = ""
-  private val IpThrottlePrefix = "ip-"
-
-  private case class ListenerQuotaEntity(listenerName: String) extends ConnectionQuotaEntity {
-    override def sensorName: String = s"$ConnectionRateSensorName-$listenerName"
-    override def sensorExpiration: Long = Long.MaxValue
-    override def metricName: String = ConnectionRateMetricName
-    override def metricTags: Map[String, String] = Map(ListenerMetricTag -> listenerName)
-  }
-
-  private case object BrokerQuotaEntity extends ConnectionQuotaEntity {
-    override def sensorName: String = ConnectionRateSensorName
-    override def sensorExpiration: Long = Long.MaxValue
-    override def metricName: String = s"broker-$ConnectionRateMetricName"
-    override def metricTags: Map[String, String] = Map.empty
-  }
-
-  private case class IpQuotaEntity(ip: InetAddress) extends ConnectionQuotaEntity {
-    override def sensorName: String = s"$ConnectionRateSensorName-${ip.getHostAddress}"
-    override def sensorExpiration: Long = InactiveSensorExpirationTimeSeconds
-    override def metricName: String = ConnectionRateMetricName
-    override def metricTags: Map[String, String] = Map(IpMetricTag -> ip.getHostAddress)
-  }
-}
-
-class ConnectionQuotas(config: KafkaConfig, time: Time, metrics: Metrics) extends Logging with AutoCloseable {
-
-  @volatile private var defaultMaxConnectionsPerIp: Int = config.maxConnectionsPerIp
-  @volatile private var maxConnectionsPerIpOverrides = config.maxConnectionsPerIpOverrides.map { case (host, count) => (InetAddress.getByName(host), count) }
-  @volatile private var brokerMaxConnections = config.maxConnections
-  private val interBrokerListenerName = config.interBrokerListenerName
-  private val counts = mutable.Map[InetAddress, Int]()
-
-  // Listener counts and configs are synchronized on `counts`
-  private val listenerCounts = mutable.Map[ListenerName, Int]()
-  private[network] val maxConnectionsPerListener = mutable.Map[ListenerName, ListenerConnectionQuota]()
-  @volatile private var totalCount = 0
-  // updates to defaultConnectionRatePerIp or connectionRatePerIp must be synchronized on `counts`
-  @volatile private var defaultConnectionRatePerIp = QuotaConfigs.IP_CONNECTION_RATE_DEFAULT.intValue()
-  private val connectionRatePerIp = new ConcurrentHashMap[InetAddress, Int]()
-  // sensor that tracks broker-wide connection creation rate and limit (quota)
-  private val brokerConnectionRateSensor = getOrCreateConnectionRateQuotaSensor(config.maxConnectionCreationRate, BrokerQuotaEntity)
-  private val maxThrottleTimeMs = TimeUnit.SECONDS.toMillis(config.quotaWindowSizeSeconds.toLong)
-
-  def inc(listenerName: ListenerName, address: InetAddress, acceptorBlockedPercentMeter: com.yammer.metrics.core.Meter): Unit = {
-    counts.synchronized {
-      waitForConnectionSlot(listenerName, acceptorBlockedPercentMeter)
-
-      recordIpConnectionMaybeThrottle(listenerName, address)
-      val count = counts.getOrElseUpdate(address, 0)
-      counts.put(address, count + 1)
-      totalCount += 1
-      if (listenerCounts.contains(listenerName)) {
-        listenerCounts.put(listenerName, listenerCounts(listenerName) + 1)
-      }
-      val max = maxConnectionsPerIpOverrides.getOrElse(address, defaultMaxConnectionsPerIp)
-      if (count >= max)
-        throw new TooManyConnectionsException(address, max)
-    }
-  }
-
-  private[network] def updateMaxConnectionsPerIp(maxConnectionsPerIp: Int): Unit = {
-    defaultMaxConnectionsPerIp = maxConnectionsPerIp
-  }
-
-  private[network] def updateMaxConnectionsPerIpOverride(overrideQuotas: Map[String, Int]): Unit = {
-    maxConnectionsPerIpOverrides = overrideQuotas.map { case (host, count) => (InetAddress.getByName(host), count) }
-  }
-
-  private[network] def updateBrokerMaxConnections(maxConnections: Int): Unit = {
-    counts.synchronized {
-      brokerMaxConnections = maxConnections
-      counts.notifyAll()
-    }
-  }
-
-  private[network] def updateBrokerMaxConnectionRate(maxConnectionRate: Int): Unit = {
-    // if there is a connection waiting on the rate throttle delay, we will let it wait the original delay even if
-    // the rate limit increases, because it is just one connection per listener and the code is simpler that way
-    updateConnectionRateQuota(maxConnectionRate, BrokerQuotaEntity)
-  }
-
-  /**
-   * Update the connection rate quota for a given IP and updates quota configs for updated IPs.
-   * If an IP is given, metric config will be updated only for the given IP, otherwise
-   * all metric configs will be checked and updated if required.
-   *
-   * @param ip ip to update or default if None
-   * @param maxConnectionRate new connection rate, or resets entity to default if None
-   */
-  def updateIpConnectionRateQuota(ip: Option[InetAddress], maxConnectionRate: Option[Int]): Unit = synchronized {
-    def isIpConnectionRateMetric(metricName: MetricName) = {
-      metricName.name == ConnectionRateMetricName &&
-      metricName.group == MetricsGroup &&
-      metricName.tags.containsKey(IpMetricTag)
-    }
-
-    def shouldUpdateQuota(metric: KafkaMetric, quotaLimit: Int) = {
-      quotaLimit != metric.config.quota.bound
-    }
-
-    ip match {
-      case Some(address) =>
-        // synchronize on counts to ensure reading an IP connection rate quota and creating a quota config is atomic
-        counts.synchronized {
-          maxConnectionRate match {
-            case Some(rate) =>
-              info(s"Updating max connection rate override for $address to $rate")
-              connectionRatePerIp.put(address, rate)
-            case None =>
-              info(s"Removing max connection rate override for $address")
-              connectionRatePerIp.remove(address)
-          }
-        }
-        updateConnectionRateQuota(connectionRateForIp(address), IpQuotaEntity(address))
-      case None =>
-        // synchronize on counts to ensure reading an IP connection rate quota and creating a quota config is atomic
-        counts.synchronized {
-          defaultConnectionRatePerIp = maxConnectionRate.getOrElse(QuotaConfigs.IP_CONNECTION_RATE_DEFAULT.intValue())
-        }
-        info(s"Updated default max IP connection rate to $defaultConnectionRatePerIp")
-        metrics.metrics.forEach { (metricName, metric) =>
-          if (isIpConnectionRateMetric(metricName)) {
-            val quota = connectionRateForIp(InetAddress.getByName(metricName.tags.get(IpMetricTag)))
-            if (shouldUpdateQuota(metric, quota)) {
-              debug(s"Updating existing connection rate quota config for ${metricName.tags} to $quota")
-              metric.config(rateQuotaMetricConfig(quota))
-            }
-          }
-        }
-    }
-  }
-
-  // Visible for testing
-  def connectionRateForIp(ip: InetAddress): Int = {
-    connectionRatePerIp.getOrDefault(ip, defaultConnectionRatePerIp)
-  }
-
-  private[network] def addListener(config: KafkaConfig, listenerName: ListenerName): Unit = {
-    counts.synchronized {
-      if (!maxConnectionsPerListener.contains(listenerName)) {
-        val newListenerQuota = new ListenerConnectionQuota(counts, listenerName)
-        maxConnectionsPerListener.put(listenerName, newListenerQuota)
-        listenerCounts.put(listenerName, 0)
-        config.addReconfigurable(newListenerQuota)
-        newListenerQuota.configure(config.valuesWithPrefixOverride(listenerName.configPrefix))
-      }
-      counts.notifyAll()
-    }
-  }
-
-  private[network] def removeListener(config: KafkaConfig, listenerName: ListenerName): Unit = {
-    counts.synchronized {
-      maxConnectionsPerListener.remove(listenerName).foreach { listenerQuota =>
-        listenerCounts.remove(listenerName)
-        // once listener is removed from maxConnectionsPerListener, no metrics will be recorded into listener's sensor
-        // so it is safe to remove sensor here
-        listenerQuota.close()
-        counts.notifyAll() // wake up any waiting acceptors to close cleanly
-        config.removeReconfigurable(listenerQuota)
-      }
-    }
-  }
-
-  def dec(listenerName: ListenerName, address: InetAddress): Unit = {
-    counts.synchronized {
-      val count = counts.getOrElse(address,
-        throw new IllegalArgumentException(s"Attempted to decrease connection count for address with no connections, address: $address"))
-      if (count == 1)
-        counts.remove(address)
-      else
-        counts.put(address, count - 1)
-
-      if (totalCount <= 0)
-        error(s"Attempted to decrease total connection count for broker with no connections")
-      totalCount -= 1
-
-      if (maxConnectionsPerListener.contains(listenerName)) {
-        val listenerCount = listenerCounts(listenerName)
-        if (listenerCount == 0)
-          error(s"Attempted to decrease connection count for listener $listenerName with no connections")
-        else
-          listenerCounts.put(listenerName, listenerCount - 1)
-      }
-      counts.notifyAll() // wake up any acceptors waiting to process a new connection since listener connection limit was reached
-    }
-  }
-
-  def get(address: InetAddress): Int = counts.synchronized {
-    counts.getOrElse(address, 0)
-  }
-
-  private def waitForConnectionSlot(listenerName: ListenerName,
-                                    acceptorBlockedPercentMeter: com.yammer.metrics.core.Meter): Unit = {
-    counts.synchronized {
-      val startThrottleTimeMs = time.milliseconds
-      val throttleTimeMs = math.max(recordConnectionAndGetThrottleTimeMs(listenerName, startThrottleTimeMs), 0)
-
-      if (throttleTimeMs > 0 || !connectionSlotAvailable(listenerName)) {
-        val startNs = time.nanoseconds
-        val endThrottleTimeMs = startThrottleTimeMs + throttleTimeMs
-        var remainingThrottleTimeMs = throttleTimeMs
-        do {
-          counts.wait(remainingThrottleTimeMs)
-          remainingThrottleTimeMs = math.max(endThrottleTimeMs - time.milliseconds, 0)
-        } while (remainingThrottleTimeMs > 0 || !connectionSlotAvailable(listenerName))
-        acceptorBlockedPercentMeter.mark(time.nanoseconds - startNs)
-      }
-    }
-  }
-
-  // This is invoked in every poll iteration and we close one LRU connection in an iteration
-  // if necessary
-  def maxConnectionsExceeded(listenerName: ListenerName): Boolean = {
-    totalCount > brokerMaxConnections && !protectedListener(listenerName)
-  }
-
-  private def connectionSlotAvailable(listenerName: ListenerName): Boolean = {
-    if (listenerCounts(listenerName) >= maxListenerConnections(listenerName))
-      false
-    else if (protectedListener(listenerName))
-      true
-    else
-      totalCount < brokerMaxConnections
-  }
-
-  private def protectedListener(listenerName: ListenerName): Boolean =
-    interBrokerListenerName == listenerName && listenerCounts.size > 1
-
-  private def maxListenerConnections(listenerName: ListenerName): Int =
-    maxConnectionsPerListener.get(listenerName).map(_.maxConnections).getOrElse(Int.MaxValue)
-
-  /**
-   * Calculates the delay needed to bring the observed connection creation rate to listener-level limit or to broker-wide
-   * limit, whichever the longest. The delay is capped to the quota window size defined by QuotaWindowSizeSecondsProp
-   *
-   * @param listenerName listener for which calculate the delay
-   * @param timeMs current time in milliseconds
-   * @return delay in milliseconds
-   */
-  private def recordConnectionAndGetThrottleTimeMs(listenerName: ListenerName, timeMs: Long): Long = {
-    def recordAndGetListenerThrottleTime(minThrottleTimeMs: Int): Int = {
-      maxConnectionsPerListener
-        .get(listenerName)
-        .map { listenerQuota =>
-          val listenerThrottleTimeMs = recordAndGetThrottleTimeMs(listenerQuota.connectionRateSensor, timeMs)
-          val throttleTimeMs = math.max(minThrottleTimeMs, listenerThrottleTimeMs)
-          // record throttle time due to hitting connection rate quota
-          if (throttleTimeMs > 0) {
-            listenerQuota.listenerConnectionRateThrottleSensor.record(throttleTimeMs.toDouble, timeMs)
-          }
-          throttleTimeMs
-        }
-        .getOrElse(0)
-    }
-
-    if (protectedListener(listenerName)) {
-      recordAndGetListenerThrottleTime(0)
-    } else {
-      val brokerThrottleTimeMs = recordAndGetThrottleTimeMs(brokerConnectionRateSensor, timeMs)
-      recordAndGetListenerThrottleTime(brokerThrottleTimeMs)
-    }
-  }
-
-  /**
-   * Record IP throttle time on the corresponding listener. To avoid over-recording listener/broker connection rate, we
-   * also un-record the listener and broker connection if the IP gets throttled.
-   *
-   * @param listenerName listener to un-record connection
-   * @param throttleMs IP throttle time to record for listener
-   * @param timeMs current time in milliseconds
-   */
-  private def updateListenerMetrics(listenerName: ListenerName, throttleMs: Long, timeMs: Long): Unit = {
-    if (!protectedListener(listenerName)) {
-      brokerConnectionRateSensor.record(-1.0, timeMs, false)
-    }
-    maxConnectionsPerListener
-      .get(listenerName)
-      .foreach { listenerQuota =>
-        listenerQuota.ipConnectionRateThrottleSensor.record(throttleMs.toDouble, timeMs)
-        listenerQuota.connectionRateSensor.record(-1.0, timeMs, false)
-      }
-  }
-
-  /**
-   * Calculates the delay needed to bring the observed connection creation rate to the IP limit.
-   * If the connection would cause an IP quota violation, un-record the connection for both IP,
-   * listener, and broker connection rate and throw a ConnectionThrottledException. Calls to
-   * this function must be performed with the counts lock to ensure that reading the IP
-   * connection rate quota and creating the sensor's metric config is atomic.
-   *
-   * @param listenerName listener to unrecord connection if throttled
-   * @param address ip address to record connection
-   */
-  private def recordIpConnectionMaybeThrottle(listenerName: ListenerName, address: InetAddress): Unit = {
-    val connectionRateQuota = connectionRateForIp(address)
-    val quotaEnabled = connectionRateQuota != QuotaConfigs.IP_CONNECTION_RATE_DEFAULT
-    if (quotaEnabled) {
-      val sensor = getOrCreateConnectionRateQuotaSensor(connectionRateQuota, IpQuotaEntity(address))
-      val timeMs = time.milliseconds
-      val throttleMs = recordAndGetThrottleTimeMs(sensor, timeMs)
-      if (throttleMs > 0) {
-        trace(s"Throttling $address for $throttleMs ms")
-        // unrecord the connection since we won't accept the connection
-        sensor.record(-1.0, timeMs, false)
-        updateListenerMetrics(listenerName, throttleMs, timeMs)
-        throw new ConnectionThrottledException(address, timeMs, throttleMs)
-      }
-    }
-  }
-
-  /**
-   * Records a new connection into a given connection acceptance rate sensor 'sensor' and returns throttle time
-   * in milliseconds if quota got violated
-   * @param sensor sensor to record connection
-   * @param timeMs current time in milliseconds
-   * @return throttle time in milliseconds if quota got violated, otherwise 0
-   */
-  private def recordAndGetThrottleTimeMs(sensor: Sensor, timeMs: Long): Int = {
-    try {
-      sensor.record(1.0, timeMs)
-      0
-    } catch {
-      case e: QuotaViolationException =>
-        val throttleTimeMs = QuotaUtils.boundedThrottleTime(e, maxThrottleTimeMs, timeMs).toInt
-        debug(s"Quota violated for sensor (${sensor.name}). Delay time: $throttleTimeMs ms")
-        throttleTimeMs
-    }
-  }
-
-  /**
-   * Creates sensor for tracking the connection creation rate and corresponding connection rate quota for a given
-   * listener or broker-wide, if listener is not provided.
-   * @param quotaLimit connection creation rate quota
-   * @param connectionQuotaEntity entity to create the sensor for
-   */
-  private def getOrCreateConnectionRateQuotaSensor(quotaLimit: Int, connectionQuotaEntity: ConnectionQuotaEntity): Sensor = {
-    Option(metrics.getSensor(connectionQuotaEntity.sensorName)).getOrElse {
-      val sensor = metrics.sensor(
-        connectionQuotaEntity.sensorName,
-        rateQuotaMetricConfig(quotaLimit),
-        connectionQuotaEntity.sensorExpiration
-      )
-      sensor.add(connectionRateMetricName(connectionQuotaEntity), new Rate, null)
-      sensor
-    }
-  }
-
-  /**
-   * Updates quota configuration for a given connection quota entity
-   */
-  private def updateConnectionRateQuota(quotaLimit: Int, connectionQuotaEntity: ConnectionQuotaEntity): Unit = {
-    Option(metrics.metric(connectionRateMetricName(connectionQuotaEntity))).foreach { metric =>
-      metric.config(rateQuotaMetricConfig(quotaLimit))
-      info(s"Updated ${connectionQuotaEntity.metricName} max connection creation rate to $quotaLimit")
-    }
-  }
-
-  private def connectionRateMetricName(connectionQuotaEntity: ConnectionQuotaEntity): MetricName = {
-    metrics.metricName(
-      connectionQuotaEntity.metricName,
-      MetricsGroup,
-      s"Tracking rate of accepting new connections (per second)",
-      connectionQuotaEntity.metricTags.asJava)
-  }
-
-  private def rateQuotaMetricConfig(quotaLimit: Int): MetricConfig = {
-    new MetricConfig()
-      .timeWindow(config.quotaWindowSizeSeconds.toLong, TimeUnit.SECONDS)
-      .samples(config.numQuotaSamples)
-      .quota(new Quota(quotaLimit, true))
-  }
-
-  def close(): Unit = {
-    metrics.removeSensor(brokerConnectionRateSensor.name)
-    maxConnectionsPerListener.values.foreach(_.close())
-  }
-
-  class ListenerConnectionQuota(lock: Object, listener: ListenerName) extends ListenerReconfigurable with AutoCloseable {
-    @volatile private var _maxConnections = Int.MaxValue
-    private[network] val connectionRateSensor = getOrCreateConnectionRateQuotaSensor(Int.MaxValue, ListenerQuotaEntity(listener.value))
-    private[network] val listenerConnectionRateThrottleSensor = createConnectionRateThrottleSensor(ListenerThrottlePrefix)
-    private[network] val ipConnectionRateThrottleSensor = createConnectionRateThrottleSensor(IpThrottlePrefix)
-
-    def maxConnections: Int = _maxConnections
-
-    override def listenerName(): ListenerName = listener
-
-    override def configure(configs: util.Map[String, _]): Unit = {
-      _maxConnections = maxConnections(configs)
-      updateConnectionRateQuota(maxConnectionCreationRate(configs), ListenerQuotaEntity(listener.value))
-    }
-
-    override def reconfigurableConfigs(): util.Set[String] = {
-      SocketServer.ListenerReconfigurableConfigs.asJava
-    }
-
-    override def validateReconfiguration(configs: util.Map[String, _]): Unit = {
-      val value = maxConnections(configs)
-      if (value <= 0)
-        throw new ConfigException(s"Invalid ${KafkaConfig.MaxConnectionsProp} $value")
-
-      val rate = maxConnectionCreationRate(configs)
-      if (rate <= 0)
-        throw new ConfigException(s"Invalid ${KafkaConfig.MaxConnectionCreationRateProp} $rate")
-    }
-
-    override def reconfigure(configs: util.Map[String, _]): Unit = {
-      lock.synchronized {
-        _maxConnections = maxConnections(configs)
-        updateConnectionRateQuota(maxConnectionCreationRate(configs), ListenerQuotaEntity(listener.value))
-        lock.notifyAll()
-      }
-    }
-
-    def close(): Unit = {
-      metrics.removeSensor(connectionRateSensor.name)
-      metrics.removeSensor(listenerConnectionRateThrottleSensor.name)
-      metrics.removeSensor(ipConnectionRateThrottleSensor.name)
-    }
-
-    private def maxConnections(configs: util.Map[String, _]): Int = {
-      Option(configs.get(KafkaConfig.MaxConnectionsProp)).map(_.toString.toInt).getOrElse(Int.MaxValue)
-    }
-
-    private def maxConnectionCreationRate(configs: util.Map[String, _]): Int = {
-      Option(configs.get(KafkaConfig.MaxConnectionCreationRateProp)).map(_.toString.toInt).getOrElse(Int.MaxValue)
-    }
-
-    /**
-     * Creates sensor for tracking the average throttle time on this listener due to hitting broker/listener connection
-     * rate or IP connection rate quota. The average is out of all throttle times > 0, which is consistent with the
-     * bandwidth and request quota throttle time metrics.
-     */
-    private def createConnectionRateThrottleSensor(throttlePrefix: String): Sensor = {
-      val sensor = metrics.sensor(s"${throttlePrefix}ConnectionRateThrottleTime-${listener.value}")
-      val metricName = metrics.metricName(s"${throttlePrefix}connection-accept-throttle-time",
-        MetricsGroup,
-        "Tracking average throttle-time, out of non-zero throttle times, per listener",
-        Map(ListenerMetricTag -> listener.value).asJava)
-      sensor.add(metricName, new Avg)
-      sensor
-    }
-  }
-
-  /**
-   * Close `channel` and decrement the connection count.
-   */
-  def closeChannel(log: Logging, listenerName: ListenerName, channel: SocketChannel): Unit = {
-    if (channel != null) {
-      log.debug(s"Closing connection from ${channel.socket.getRemoteSocketAddress}")
-      dec(listenerName, channel.socket.getInetAddress)
-      closeSocket(channel, log)
-    }
-  }
-
 }
 
 class TooManyConnectionsException(val ip: InetAddress, val count: Int) extends KafkaException(s"Too many connections from $ip (maximum = $count)")
